[UPD] write mem

+++ include/pmap.hpp
@@ -40,7 +40,7 @@ public:
     int map_pid(pid_t __pid);
     bool map_mem(std::string __mem);
 
-    bool map_write(off_t, void *, uint8_t);
+    bool map_write(off_t, void *, uint);
     bool map_read(off_t, uint8_t, Data &);
     int map_find(off_t, uint64_t, std::string, uint8_t, std::vector<off_t> &);
     void map_stop(bool = true);
